# src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: phenix API
  version: 0.0.0
  title: phenix
  termsOfService: http://swagger.io/terms/
  contact:
    url: https://github.com/sandialabs/sceptre-phenix
  license:
    name: GNU General Public License v3.0
...
paths:
...
  "/disks":
    get:
      tags:
        - Disks
      summary: Get all available disk images
      description: >-
        Lists the disk images in the minimega files directory and its folders,
        the images the experiments' topologies use wherever they are, and the
        images backing any of them, each once by its full path. The folders
        searched go at most 16 levels below the files directory and leave out
        hidden names (starting with `.`), `lost+found`, the experiment folders
        `<experiment>/files`, `<experiment>/tmp` and
        `<experiment>/miniccc_responses`, and the top-level `saved` and
        `transfer_*` folders; symlinked folders are not followed. Needs the
        `disks` `list` permission. A `disks` name is an image's file name, so
        it covers the files of that name in every folder. Lists only the images
        whose names the role may list, and of those outside the files
        directory only the ones an experiment the role may `get` uses, or that
        back an image the role sees. Each image's experiments are cut to those
        the role may `list`.
      operationId: getDisks
      parameters:
        - name: expName
          in: query
          description: >-
            look only in this experiment's topology, rather than every
            experiment's, for the images and uses it reports; needs the
            `experiments` `get` permission for it
          required: false
          schema:
            type: string
        - name: diskType
          in: query
          description: comma-separated kinds to list (vm, container, iso, unknown); all by default
          required: false
          schema:
            type: string
        - name: refresh
          in: query
          description: >-
            set to true to inspect every image again rather than reuse what is
            known about images that have not changed
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disks"
        "403":
          description: >-
            forbidden, including an `expName` the role may not get
        "404":
          description: no experiment named `expName`
    post:
      tags:
        - Disks
      summary: Upload a disk image
      description: >-
        Writes the uploaded file to the top level of the minimega files
        directory, under the file's own name. Needs the `disks` `upload`
        permission.
      operationId: postDisks
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: successful operation
        "403":
          description: forbidden
        "500":
          description: the upload could not be read or written
    delete:
      tags:
        - Disks
      summary: Delete a disk image
      description: >-
        Deletes the image's file. Images backed by it stop working. Needs the
        `disks` `delete` permission for the image's file name.
      operationId: deleteDisks
      parameters:
        - name: disk
          in: query
          description: >-
            the image, relative to the minimega files directory or absolute;
            it must be inside that directory and not in a folder the disk list
            leaves out
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "400":
          description: >-
            a path outside the minimega files directory or in a folder the
            disk list leaves out, or a folder
        "403":
          description: forbidden
        "404":
          description: no such image
        "500":
          description: the file could not be deleted
...
servers:
  - url: http://localhost:3000/api/v1
components:
  schemas:
...
    Disks:
      type: object
      properties:
        disks:
          type: array
          description: sorted by name, then by full path
          items:
            $ref: "#/components/schemas/Disk"
    Disk:
      type: object
      properties:
        kind:
          type: string
          enum: [VM, Container, ISO, Unknown]
        name:
          type: string
          description: >-
            the image's file name, which `disks` permissions name; images in
            different folders can share it
        fullPath:
          type: string
          description: >-
            the image's absolute path, without `.` or `..` segments, which
            identifies it
        relativePath:
          type: string
          description: >-
            the image's path relative to the minimega files directory, or empty
            when the image is outside it
        outsideFilesDir:
          type: boolean
          description: whether the image is outside the minimega files directory
        readOnly:
          type: boolean
          description: >-
            whether the disk actions refuse the image: it is outside the
            minimega files directory, or in a folder the disk list leaves out,
            such as an experiment's files
        size:
          type: string
          description: size of the file, as a readable string
        virtualSize:
          type: string
          description: size the guest sees, as a readable string
        experiments:
          type: array
          description: >-
            experiments whose topology uses the image, directly or as the
            backing image of an image it uses
          nullable: true
          items:
            type: object
            properties:
              name:
                type: string
              running:
                type: boolean
        backingImages:
          type: array
          description: >-
            the full paths of the images this one is backed by, nearest first
          items:
            type: string
        inUse:
          type: boolean
          description: whether the image's file is in use, such as by a running VM
...